# Optimize pairLabelsAndConfidence to reduce memory allocations