… internal/birdnet/analyze_test.go