## Summary
- Optimizes the `pairLabelsAndConfidence` function to eliminate memory allocations during bird species predictions
- Addresses memory usage shown in pprof analysis where this function consumed 14.72MB (22% of heap)
- Implements buffer reuse pattern to achieve zero-allocation operation

## Changes
1. Added `resultsBuffer []datastore.Results` field to BirdNET struct
2. Created `pairLabelsAndConfidenceReuse` function that updates buffer in-place
3. Pre-allocate buffer during model validation
4. Updated `Predict` function to use the optimized version
5. Added comprehensive tests to ensure functional correctness
6. Added benchmarks to validate performance improvements

## Performance Impact
Based on benchmarks with 6,522 species (BirdNET v2.4 size):

**Memory Allocations:**
- Before: 262,150 bytes per prediction
- After: 0 bytes per prediction
- **100% reduction in allocations**

**Execution Speed:**
- Before: ~180,000 ns/op
- After: ~6,000 ns/op  
- **30x faster execution**

**Production Impact:**
- For a system processing 10 predictions/second per audio source
- Saves ~2.5MB/second of allocations
- Reduces GC pressure significantly
- Should eliminate ~14.72MB from heap usage

## Test Results
All tests pass, including:
- Correctness tests for various edge cases
- Buffer reuse tests
- Existing BirdNET tests remain passing

## Benchmark Results
```
BenchmarkPairLabelsAndConfidence-16             	  107370	    179721 ns/op	  262150 B/op	       1 allocs/op
BenchmarkPairLabelsAndConfidenceOptimized-16    	 1979150	      6061 ns/op	       0 B/op	       0 allocs/op
```

🤖 Generated with [Claude Code](https://claude.ai/code)